remove the filterbank_data concept and just stick with scan
