+++ bliss/core/include/core/pyblisscore.hpp
@@ -2,7 +2,6 @@
 
 #include "cadence.hpp"
 #include "coarse_channel.hpp"
-#include "filterbank_data.hpp"
 #include "scan.hpp"
 
 #include <nanobind/nanobind.h>
@@ -41,29 +40,31 @@ void bind_pycore(nb::module_ m) {
             .def_ro("noise_estimate", &bliss::coarse_channel::_noise_stats)
                 ;
 
-    nb::class_<bliss::filterbank_data>(m, "filterbank_data")
-            .def(nb::init<std::string>())
-            .def("get_coarse_channel", &bliss::filterbank_data::get_coarse_channel)
-            .def("get_channelization", &bliss::filterbank_data::get_channelization)
-            .def("get_coarse_channel_with_frequency", &bliss::filterbank_data::get_coarse_channel_with_frequency)
-            .def_prop_ro("num_coarse_channels", &bliss::filterbank_data::get_number_coarse_channels)
-            .def_prop_ro("az_start", &bliss::filterbank_data::az_start)
-            .def_prop_ro("data_type",&bliss::filterbank_data::data_type)
-            .def_prop_ro("fch1", &bliss::filterbank_data::fch1)
-            .def_prop_ro("foff", &bliss::filterbank_data::foff)
-            .def_prop_ro("machine_id", &bliss::filterbank_data::machine_id)
-            .def_prop_ro("nbits", &bliss::filterbank_data::nbits)
-            .def_prop_ro("nchans", &bliss::filterbank_data::nchans)
-            .def_prop_ro("source_name", &bliss::filterbank_data::source_name)
-            .def_prop_ro("src_dej", &bliss::filterbank_data::src_dej)
-            .def_prop_ro("src_raj", &bliss::filterbank_data::src_raj)
-            .def_prop_ro("telescope_id", &bliss::filterbank_data::telescope_id)
-            .def_prop_ro("tsamp", &bliss::filterbank_data::tsamp)
-            .def_prop_ro("tstart", &bliss::filterbank_data::tstart)
-            .def_prop_ro("za_start", &bliss::filterbank_data::za_start)
-        //     .def("__getstate__", &bliss::filterbank_data::get_state<false>)
+    nb::class_<bliss::scan>(m, "scan")
+            .def(nb::init<std::string_view>(), "file_path"_a)
+            .def("get_coarse_channel", &bliss::scan::get_coarse_channel)
+            .def("get_channelization", &bliss::scan::get_channelization)
+            .def("get_coarse_channel_with_frequency", &bliss::scan::get_coarse_channel_with_frequency, "frequency"_a);
+            .def("extract_coarse_channels", &bliss::scan::extract_coarse_channels, "start"_a=0, "count"_a=1)
+            .def("hits", &bliss::scan::hits)
+            .def_prop_ro("num_coarse_channels", &bliss::scan::get_number_coarse_channels)
+            .def_prop_ro("az_start", &bliss::scan::az_start)
+            .def_prop_ro("data_type",&bliss::scan::data_type)
+            .def_prop_ro("fch1", &bliss::scan::fch1)
+            .def_prop_ro("foff", &bliss::scan::foff)
+            .def_prop_ro("machine_id", &bliss::scan::machine_id)
+            .def_prop_ro("nbits", &bliss::scan::nbits)
+            .def_prop_ro("nchans", &bliss::scan::nchans)
+            .def_prop_ro("source_name", &bliss::scan::source_name)
+            .def_prop_ro("src_dej", &bliss::scan::src_dej)
+            .def_prop_ro("src_raj", &bliss::scan::src_raj)
+            .def_prop_ro("telescope_id", &bliss::scan::telescope_id)
+            .def_prop_ro("tsamp", &bliss::scan::tsamp)
+            .def_prop_ro("tstart", &bliss::scan::tstart)
+            .def_prop_ro("za_start", &bliss::scan::za_start)
+        //     .def("__getstate__", &bliss::scan::get_state<false>)
         //     .def("__setstate__",
-        //          [](bliss::filterbank_data   &self,
+        //          [](bliss::scan   &self,
         //             const std::tuple<bland::ndarray,
         //                              bland::ndarray,
         //                              double,
@@ -81,7 +82,7 @@ void bind_pycore(nb::module_ m) {
         //                              int64_t,
         //                              double,
         //                              double> &state) {
-        //              new (&self) bliss::filterbank_data(std::get<0>(state),
+        //              new (&self) bliss::scan(std::get<0>(state),
         //                                                 std::get<1>(state),
         //                                                 std::get<2>(state),
         //                                                 std::get<3>(state),
@@ -134,20 +135,8 @@ void bind_pycore(nb::module_ m) {
             .def_rw("magnitude", &bliss::integrated_flags::magnitude)
             .def_rw("sigma_clip", &bliss::integrated_flags::sigma_clip);
 
-    nb::class_<bliss::scan, bliss::filterbank_data>(m, "scan")
-            .def(nb::init<const bliss::filterbank_data &>())
-            .def(nb::init<std::string_view>(), "file_path"_a)
-            .def("extract_corase_channels", &bliss::scan::extract_coarse_channels, "start"_a=0, "count"_a=1)
-            .def("hits", &bliss::scan::hits)
-            //    .def("__getstate__", &bliss::scan::get_state)
-            //    .def("__setstate__", [](bliss::scan &self, const std::tuple<bliss::noise_stats, int,
-            //    std::vector<bliss::hit>, int64_t> &state_tuple) {
-            //         new (&self) bliss::scan();
-            //    })
-            ;
-
     nb::class_<bliss::observation_target>(m, "observation_target")
-            .def(nb::init<std::vector<bliss::filterbank_data>>())
+            .def(nb::init<std::vector<bliss::scan>>())
             .def(nb::init<std::vector<std::string_view>>())
             .def_rw("scans", &bliss::observation_target::_scans)
             .def_rw("target_name", &bliss::observation_target::_target_name);
