+++ bliss/flaggers/include/flaggers/magnitude.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <core/filterbank_data.hpp>
+#include <core/scan.hpp>
 #include <core/cadence.hpp>
 
 namespace bliss {
@@ -20,18 +20,18 @@ coarse_channel flag_magnitude(coarse_channel fb_data, float threshold);
 coarse_channel flag_magnitude(coarse_channel fb_data);
 
 /**
- * return a masked copy of fb_data where the filterbank_data.data() is above the given threshold
+ * return a masked copy of fb_data where the scan.data() is above the given threshold
  * 
 */
-filterbank_data flag_magnitude(filterbank_data fb_data, float threshold);
+scan flag_magnitude(scan fb_data, float threshold);
 
 /**
- * return a masked copy of fb_data where the filterbank_data.data() is above a threshold.
+ * return a masked copy of fb_data where the scan.data() is above a threshold.
  * When no threshold is given, this will internally compute a mean & stddev, then use a threshold
  * of mean + 10 * stddev
  * 
 */
-filterbank_data flag_magnitude(filterbank_data fb_data);
+scan flag_magnitude(scan fb_data);
 
 observation_target flag_magnitude(observation_target observations, float threshold);
 observation_target flag_magnitude(observation_target observations);
