+++ bliss/flaggers/magnitude.cpp
@@ -37,7 +37,7 @@ coarse_channel bliss::flag_magnitude(coarse_channel cc_data) {
  * return a mask with flags set for grid elements which have very large magnitudes
  * 
 */
-filterbank_data bliss::flag_magnitude(filterbank_data fil_data, float threshold) {
+scan bliss::flag_magnitude(scan fil_data, float threshold) {
     auto number_coarse_channels = fil_data.get_number_coarse_channels();
     for (auto cc_index = 0; cc_index < number_coarse_channels; ++cc_index) {
         auto cc = fil_data.get_coarse_channel(cc_index);
@@ -46,7 +46,7 @@ filterbank_data bliss::flag_magnitude(filterbank_data fil_data, float threshold)
     return fil_data;
 }
 
-filterbank_data bliss::flag_magnitude(filterbank_data fil_data) {
+scan bliss::flag_magnitude(scan fil_data) {
     auto number_coarse_channels = fil_data.get_number_coarse_channels();
     for (auto cc_index = 0; cc_index < number_coarse_channels; ++cc_index) {
         auto cc = fil_data.get_coarse_channel(cc_index);
